refactor: centralize supabase interactions
